storage,kv: Update PreIngestDelay to account for L0 sublevels
This change updates PreIngestDelay calculation to account for
overlapping L0 SSTs as opposed to all L0 SSTs. This matches
the new compaction scoring logic in Pebble, as well as the new
definitions of L0{Compaction,StopWrites}Threshold.

Also, a new Pebble option, FlushSplitBytes, is now modified in
the metamorphic test.

Release note: None.